+++ src/Domain/ValueObject/UserId.ts
@@ -1,0 +1,5 @@
+import { StringValue } from "./StringValue";
+import { ValueObject } from "./ValueObject";
+
+@ValueObject('Account')
+export class UserId extends StringValue {}